mutation authenticated

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -85,7 +85,15 @@ const Mutation = {
       }
     }, info);
   },
-  async updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+    if(!postExists) throw new Error('Unable to update post')
     return prisma.mutation.updatePost({
       where: {
         id: args.id
@@ -93,13 +101,14 @@ const Mutation = {
       data: args.data
     }, info);
   },
-  async createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -110,14 +119,30 @@ const Mutation = {
       }
     }, info);
   },
-  async deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if(!commentExists) throw new Error('Unable to delete comment');
     return prisma.mutation.deleteComment({
       where: {
         id: args.id
       }
     }, info);
   },
-  async updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if(!commentExists) throw new Error('Unable to update comment');
     return prisma.mutation.updateComment({
       where: {
         id: args.id
